Remove the old api folder
The api tests are quite neglected - the tests work for 32-bit i386 only
and there hasn't been any change in this folder since more than 2.5 years.
Additionally, there is nowadays another way of testing the KVM API - the
KVM selftests (as part of the Linux kernel sources) have a much higher
traction and feature much more tests already, so it's unlikely that the
API tests in the kvm-unit-tests repository will get much more attention
in the future. Thus let's delete the api folder now to remove the burder
from the kvm-unit-test maintainers of dragging this code along.
If someone still wants to run the dirty-log-perf test for example, they
can check out an older state of the repository (and then e.g. port the
code to the KVM selftests framework instead).

Signed-off-by: Thomas Huth <thuth@redhat.com>
Acked-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>