nVMX: Pass exit reason union to v1 exit handlers
Pass the recently introduce "union exit_reason" to the v1 exit handlers
and use it in lieu of a manual VMREAD of the exit reason.

Opportunistically fix a variety of warts in the handlers, e.g. grabbing
only bits 7:0 of the exit reason.  Modify the "Unknown exit reason"
prints to display the exit reason in hex format to make a failed
VM-Entry more recognizable.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>