x86: vmx: Fix "EPT violation - paging structure" test
Running the tests with more than 4GB of memory results in the following
failure:

  FAIL: EPT violation - paging structure

It appears that the test mistakenly used get_ept_pte() to retrieve the
guest PTE, but this function is intended for accessing EPT and not the
guest page tables. Use get_pte_level() instead of get_ept_pte().

Tested on bare-metal only.

Signed-off-by: Nadav Amit <namit@vmware.com>
Message-Id: <20200321050555.4212-1-namit@vmware.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>