arch-run: Add reserved variables to the default environ
Add the already reserved (see README) variables to the default
environ. To do so neatly we rework the environ creation a bit too.
mkstandalone also learns to honor config.mak as to whether or not
to make environs, and we allow the $ERRATATXT file to be selected
at configure time.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Message-Id: <20200407113312.65587-1-drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>