x86: svm: avoid advancing rip incorrectly on exc_inject
exc_inject advances the ripon every stage, so it can do so 3 times, but
there are only 2 vmmcall instructions that the guest runs. So, if a
failure happens on the last test, there is no vmmcall instruction to
trigger an exit.

Advance the rip only in the two stages in which vmmcall is expected to
run.

Signed-off-by: Nadav Amit <namit@vmware.com>
Message-Id: <20200630094516.22983-6-namit@vmware.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>