gitlab-ci.yml: Extend the lists of tests that we run with TCG
Thank to the recent fixes, there are now quite a lot of additional 32-bit
x86 tests that we can run in the CI.
And thanks to the update to Fedora 32 (that introduced a newer version of
QEMU), there are now also some additional tests that we can run with TCG
for the other architectures.
Note that for arm/aarch64, we now also set the MAX_SMP to be able to run
SMP-tests with TCG in the single-threaded CI containers, too.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20200701100615.7975-1-thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>