svm: Add ability to execute test via test_run on a vcpu other than vcpu 0
When running tests that can result in a vcpu being left in an
indeterminate state it is useful to be able to run the test on
a vcpu other than 0. This patch allows test_run to be executed
on any vcpu indicated by the on_vcpu member of the svm_test struct.
The initialized state of the vcpu0 registers used to populate the
vmcb is carried forward to the other vcpus.

Signed-off-by: Cathy Avery <cavery@redhat.com>
Message-Id: <20200717113422.19575-2-cavery@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>