nSVM: Add a test for NPF due to P=0 in NPT entry
If the P (present) bit in an NPT entry is cleared, accesses to the
corresponding guest physical address will fail and the guest will exit to
the host with an exit code of 0x400 (#NPF). The test checks for bits 2
and 32 to be set (respectively because all NPT accesses are considered
user accesses, and because the test encounters the P=0 entry while
translating the final guest physical address and not a guest page table.

Signed-off-by: Krish Sadhukhan <krish.sadhukhan@oracle.com>
Message-Id: <20200829005720.5325-1-krish.sadhukhan@oracle.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>