travis.yml: Change matrix keyword to jobs
jobs keyword is used throughout Travis documentation and matrix is an
alias for it (according to Travis config validation):

  root: key matrix is an alias for jobs, using jobs

At first glance it's not clear if they're the same and if jobs
documentation applies to matrix. Changing keyword name should make it
obvious.

While at it, fix the Travis config warning:

  root: deprecated key sudo (The key `sudo` has no effect anymore.)

Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>
Message-Id: <20200901085056.33391-10-r.bolshakov@yadro.com>
Tested-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>