configure: Add a check for the bash version
Our scripts do not work with older versions of the bash, like the
default Bash 3 from macOS (e.g. we use the "|&" operator which has
been introduced in Bash 4.0). Add a check to make sure that we use
at least version 4 to avoid that the users run into problems later.

Message-Id: <20200925143852.227908-1-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>