travis.yml: Rework the x86 64-bit tests
We currently have three test jobs here:

1) gcc, in-tree build
2) gcc, out-of-tree build
3) clang, in-tree build

Keeping everything in perspective, it should be sufficient to only use two
build jobs for this, one in-tree with one compiler, and one out-of-tree
with the other compiler.
So let's re-order the jobs accordingly now. And while we're at it, make
sure that all additional tests that work with the newer QEMU from Ubuntu
Focal now are tested, too, and that we check all possible tests with
Clang.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20201001072234.143703-2-thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>